llamaindex_agentic_ai_poc/llamaindex_ppe_agent+workflow_server/requirements.txt
llama_index.embeddings.huggingface==0.6.1
llama-index-llms-ollama==0.9.0
llama-index==0.14.8
llama_index.tools.mcp==0.4.3
llama-index-vector-stores-postgres==0.7.1
llama-index-workflows==2.11.1
pydantic
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastuuid==0.14.0
filelock==3.20.0
filetype==1.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
griffe==1.15.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
litellm==1.79.3
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.8
llama-index-cli==0.5.3
llama-index-core==0.14.8
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-litellm==0.6.3
llama-index-llms-ollama==0.9.0
llama-index-llms-openai==0.6.8
llama-index-readers-file==0.5.4
llama-index-readers-llama-parse==0.5.1
llama-index-tools-mcp==0.4.3
llama-index-vector-stores-postgres==0.7.1
llama-index-workflows==2.11.1
llama-parse==0.6.54
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mcp==1.21.0
mdformat==0.7.22
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.2.6
ollama==0.6.0
openai==2.7.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.2.3
pgvector==0.4.1
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
propcache==0.4.1
psutil==7.1.3
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==13.9.4
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
step==0.1.0
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typer==0.9.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.227
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
